requirements.txt
# Core dependencies
httpx>=0.25.0
requests>=2.31.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0

# Data processing
pandas>=2.1.0
pydantic>=2.4.0
pydantic-settings>=2.0.0

# Database (will add when ready)
# psycopg[binary]>=3.1.0
# asyncpg>=0.29.0
# sqlalchemy>=2.0.0
# alembic>=1.12.0

# Database - Active (PostgreSQL + pgvector)
sqlalchemy>=2.0.0
alembic>=1.12.0
asyncpg>=0.29.0  # For production PostgreSQL async
psycopg[binary]>=3.1.0  # For Alembic migrations and sync utilities
pgvector>=0.2.4  # Vector similarity search

# Orchestration
prefect==3.4.24  # Match production Prefect worker image
prefect-sqlalchemy>=0.4.0

# Redis (active - for Prefect result backend)
redis>=5.0.0

# Storage & Streaming
minio>=7.2.0  # S3-compatible object storage
aiokafka>=0.10.0  # Kafka async client
tenacity>=8.2.2  # Retry utility used by adapters

# CLI & Reporting
rich>=13.7.0  # Beautiful terminal output
feedgen>=0.9.0  # RSS/Atom feed generation

# API - ACTIVE (REST and GraphQL)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
strawberry-graphql>=0.209.0  # GraphQL schema served in production

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development
ruff>=0.1.0
black>=23.10.0
mypy>=1.6.0
